#5 moved loginHelper to helpers folder
